Moving desktop DI configuration to :desktopApp
