+++ gradle/libs.versions.toml
@@ -35,6 +35,7 @@ circuit-runtime-presenter = { module = "com.slack.circuit:circuit-runtime-presen
 circuit-retained = { module = "com.slack.circuit:circuit-retained", version.ref = "circuit" }
 circuitx-gesturenav = { module = "com.slack.circuit:circuitx-gesture-navigation", version.ref = "circuit" }
 androidx-material3-android = { group = "androidx.compose.material3", name = "material3-android", version.ref = "material3Android" }
+uuid = "com.benasher44:uuid:0.8.4"
 
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
