+++ sample/desktopApp/src/main/kotlin/com/r0adkll/kimchi/restaurant/di/DesktopApplicationComponent.kt
@@ -1,19 +1,17 @@
 // Copyright (C) 2023 r0adkll
 // SPDX-License-Identifier: Apache-2.0
-package com.r0adkll.kimchi.restaurant
+package com.r0adkll.kimchi.restaurant.di
 
 import androidx.compose.ui.unit.Density
 import com.r0adkll.kimchi.annotations.MergeComponent
 import com.r0adkll.kimchi.restaurant.common.scopes.AppScope
 import com.r0adkll.kimchi.restaurant.common.scopes.SingleIn
-import com.r0adkll.kimchi.restaurant.di.SharedAppComponent
 import java.util.prefs.Preferences
 import me.tatarka.inject.annotations.Provides
 
 @SingleIn(AppScope::class)
 @MergeComponent(
   scope = AppScope::class,
-  excludes = [WindowComponent::class],
 )
 abstract class DesktopApplicationComponent : SharedAppComponent {
